--- src/test/java/AttendanceBookTest.java
@@ -1,35 +1,0 @@
-import static constants.TestTimeMaker.MONDAY_ATTEND;
-import static constants.TestTimeMaker.NON_OPERATING_TIME;
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
-
-import domain.AttendanceBook;
-import domain.Crew;
-import java.time.LocalDate;
-import java.util.Map;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-class AttendanceBookTest {
-
-    @Test
-    @DisplayName("출석부에_기존이름이_존재여부_확인")
-    void 출석부에_기존이름이_존재여부_확인() {
-        AttendanceBook attendanceBook = new AttendanceBook();
-        Crew crew1 = Crew.createByName("쿠키");
-        crew1.addDailyAttendance(Map.of(LocalDate.of(2024, 12, 2), MONDAY_ATTEND));
-        attendanceBook.addNewCrew(crew1);
-
-        assertThat(attendanceBook.checkCrewAlreadyExists("쿠키")).isEqualTo(true);
-        assertThat(attendanceBook.checkCrewAlreadyExists("없음")).isEqualTo(false);
-    }
-
-    @Test
-    void 캠퍼스_운영_시간이_아닌_경우_예외를_출력한다() {
-        AttendanceBook attendanceBook = new AttendanceBook();
-        assertThatThrownBy(
-                () -> attendanceBook.validateIsInOperationHour(NON_OPERATING_TIME))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("[ERROR] 캠퍼스 운영 시간은 08:00~23:00 입니다.");
-    }
-}
